:sparkles: map story.author, video.viewCount

+++ app/src/main/kotlin/com/bguina/eurosport/test/data/http/model/VideoDto.kt
@@ -16,5 +16,5 @@ data class VideoDto(
     @Json(name = "url")
     val url: String? = null,
     @Json(name = "views")
-    val views: Int? = null
+    val views: Long? = null
 )
